removed product ID on datagridview in products form
PK was incrementing automatically, so when a new product was added, it would increment up one, but if you removed it, the next product added would be one above the original.  I just removed the ID from the view for now.  eg.  Add product - PK ID = 1, then delete it.  Add a new product, PK ID = 2 which is correct but client might be confused as from a user standpoint would be better to delete the old primary key.  I just removed it from view